+++ backend/src/main/java/org/cosmic/backend/domain/playList/repositorys/AlbumRepository.java
@@ -4,10 +4,12 @@ import org.cosmic.backend.domain.favoriteArtist.dtos.ArtistDetail;
 import org.cosmic.backend.domain.playList.domains.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public interface AlbumRepository extends JpaRepository<Album,Long> {
     List<Album> findAllByArtist_ArtistId(Long artistId);
     List<Album> findAllByArtist_ArtistName(String artistName);
