+++ backend/src/main/java/org/cosmic/backend/domain/playList/repositorys/TrackRepository.java
@@ -2,10 +2,12 @@ package org.cosmic.backend.domain.playList.repositorys;
 
 import org.cosmic.backend.domain.playList.domains.Track;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public interface TrackRepository extends JpaRepository<Track,Long> {
     Optional<Track> findByTitle(String title);
     Optional<Track> findByTitleAndAlbum_AlbumId(String title, Long albumId);
