+++ backend/src/main/java/org/cosmic/backend/domain/search/apis/SpotifySearchApi.java
@@ -5,9 +5,10 @@ import io.swagger.v3.oas.annotations.Operation;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.cosmic.backend.domain.auth.applications.CreateSpotifyToken;
-import org.cosmic.backend.domain.search.applications.SearchApplication;
+import org.cosmic.backend.domain.search.applications.SearchService;
 import org.cosmic.backend.domain.search.dtos.SearchRequest;
 import org.cosmic.backend.globals.annotations.ApiCommonResponses;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,10 +17,9 @@ import org.springframework.web.bind.annotation.*;
 @ApiCommonResponses
 @Tag(name = "스포티파이 검색 관련 API", description = "스포티파이를 통한 데이터 검색 제공")
 public class SpotifySearchApi {
-
+    @Autowired
+    private SearchService searchService;
     CreateSpotifyToken createSpotifyToken=new CreateSpotifyToken();
-    SearchApplication searchApplication=new SearchApplication();
-
     @GetMapping("/spotifyToken")
     public ResponseEntity<String> getToken(){ //q는 검색어
         return ResponseEntity.ok(createSpotifyToken.accesstoken());
@@ -29,7 +29,18 @@ public class SpotifySearchApi {
     @GetMapping("/searchSpotify/track")
     @Operation(summary = "특정 노래 조회")
     public ResponseEntity<String> searchTrack(@RequestBody SearchRequest searchRequest) throws JsonProcessingException { //q는 검색어
+        return ResponseEntity.ok(searchService.searchTrack(searchRequest.getAccessToken(),searchRequest.getQ()));
+    }
 
-        return ResponseEntity.ok(searchApplication.searchTrack(searchRequest.getAccessToken(),searchRequest.getQ()));
+    @GetMapping("/searchSpotify/artist")
+    @Operation(summary = "특정 아티스트 조회")
+    public ResponseEntity<String> searchArtist(@RequestBody SearchRequest searchRequest) throws JsonProcessingException { //q는 검색어
+        return ResponseEntity.ok(searchService.searchArtist(searchRequest.getAccessToken(),searchRequest.getQ()));
+    }
+
+    @GetMapping("/searchSpotify/album")
+    @Operation(summary = "특정 앨범 조회")
+    public ResponseEntity<String> searchAlbum(@RequestBody SearchRequest searchRequest) throws JsonProcessingException { //q는 검색어
+        return ResponseEntity.ok(searchService.searchAlbum(searchRequest.getAccessToken(),searchRequest.getQ()));
     }
 }
