+++ backend/src/main/java/org/cosmic/backend/domain/search/applications/SearchArtistService.java
@@ -1,0 +1,4 @@
+package org.cosmic.backend.domain.search.applications;
+
+public class SearchArtistService {
+}
